refactor: 페이지 반환과 api 반환 컨트롤러 분리
