+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -7,34 +7,26 @@ import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.entity.Reservation;
 
-@Controller
+@RestController
 public class ReservationController {
 
     private List<Reservation> reservations = new ArrayList<>();
     private AtomicLong index = new AtomicLong(1);
 
-    @GetMapping("/admin/reservation")
-    public String reservation(
-    ) {
-        return "/admin/reservation-legacy.html";
-    }
-
     @GetMapping("/reservations")
-    @ResponseBody
     public List<Reservation> getReservations(
     ) {
         return reservations;
     }
 
     @PostMapping("/reservations")
-    @ResponseBody
     public ResponseEntity<Reservation> create(
             String name,
             LocalDate date,
