+++ src/main/java/roomescape/page/AdminPageController.java
@@ -1,11 +1,10 @@
-package roomescape;
+package roomescape.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @Controller
-public class RoomescapeController {
+public class AdminPageController {
 
     @GetMapping("/admin")
     public String admin(
